refactor[RESOLVER]: Modularize custom scalars and utils function
